+++ Day5/groupApp/src/Navbar/style.css
@@ -16,7 +16,7 @@
   backdrop-filter: blur(30px);
   padding: 30px 30px;
   position: fixed;
-  top: 10;
+  top:10;
   box-shadow: rgba(50, 50, 93, 0.25) 0px 6px 12px -2px, rgba(0, 0, 0, 0.3) 0px 3px 7px -3px;
 }
 
@@ -34,31 +34,7 @@
   font-weight: 900;
   font-size: 20px;
 }
-.searchBar {
-  border: 0px;
-  width: 300px;
-  height: 35px;
-  padding-left: 20px;
-  flex-grow: 1;
-}
-.searchBar:focus {
-  outline: none;
-}
-.searchBarDiv {
-  display: flex;
-  align-items: center;
-  padding: 5px;
-  border-radius: 18px;
-  width: 350px;
-  height: 35px;
-  background-color: rgb(255, 255, 255);
-}
-.searchIcon {
-  width: 5%;
-  height: 20px;
-  padding-right: 14px;
-  cursor: pointer;
-}
+
 .avatar {
   height: 35px;
   width: 35px;
