+++ Day5/groupApp/src/style.css
@@ -26,5 +26,43 @@
   flex-wrap: wrap;
   justify-content: center;
   align-items: flex-start;
+  top: 20px;
+}
+.bottomDiv1{
   margin-top: 130px;
+  top: 50;
+}
+.searchBar {
+  border: 0px;
+  width: 300px;
+  height: 35px;
+  padding-left: 20px;
+  flex-grow: 1;
+  background-color: #EAEAEA;
+}
+.searchBar:focus {
+  outline: none;
+}
+.searchBarDiv {
+  display: flex;
+  align-items: center;
+  padding: 5px;
+  border-radius: 18px;
+  width: 350px;
+  height: 35px;
+  background-color: #EAEAEA;
+}
+.searchIcon {
+  width: 45%;
+  height: 20px;
+  padding-right: 14px;
+  cursor: pointer;
+}
+.buttonSearch{
+  background-color: transparent;
+  border: 0px;
+}
+.searchBarCont{
+  display: flex;
+  justify-content: center;
 }
